+++ coloraide/spaces/srgb/css.py
@@ -16,7 +16,6 @@ RE_COMPRESS = re.compile(r'(?i)^#({hex})\1({hex})\2({hex})\3(?:({hex})\4)?$'.for
 class SRGB(base.SRGB):
     """SRGB class."""
 
-    START = re.compile(r'(?i)\brgba?\(')
     MATCH = re.compile(
         r"""(?xi)
         (?: